fix(site): replace unnecessary non-breaking spaces
Non-breaking spaces need only be used when a space needs to be put
between two elements and you don't want the browser to break the text
across multiple lines at that point where the elements meet.

This is rarely needed, and most non-breaking spaces in the site content
weren't necessary. A couple in the stories for the header component were
relevant and retained.

on-behalf-of: @ithaka <dane.hillard@ithaka.org>