fix(chromatic): update delays for visual snapshot tests (#359)
The stories changed here have delays built into them using `setTimeout`,
but they `delay` value for Chromatic was either unset or set to the same
value as the timeout. This results regularly in snapshots where the
component state has not yet reached the desired one for testing, leading
to noisy snapshot diffs on each pull request.

Updating the `delay` value to be longer than the `setTimeout` values
should result in a more stable testing experience.