fix(toggle-button): indicate toggle button state on underlying button elements
Because the accessibility object model (AOM) is still in active
development and doesn't enjoy wide browser vendor adoption at present,
we need to maintain the accessibility for assistive technologies in the
underlying native elements rather than the custom element host
containers. For things like buttons, this often means adding a property
that gets rendered onto the underlying native element as an `aria-`
attribute.