feat(storybook): enable new build options
This necessitated getting rid of dynamic component titles, namely for
showing e.g. "Pharos 12.13.0" in the left nav. I opted instead to inject
the version as the title of the intro page. I had to wrestle with MDX a
bit to do this; it appears the "components" and "props" features
documented on their site are v2 features, but our Storybook is still
using v1. I attemped to enable the v2 preview, but HTML comments cause a
parsing error currently and we need them for our README setup to
function properly.