fix(tabs): fix initial tab selection in more cases
The design flaw in tabs was letting each individual tab manage its own
state when clicked. This created issues with state management
centralization and reasoning. This change instead moves the job of state
management to the Tabs component, which coordinates the individual Tab
and TabPanel components it contains. This results in smoother management
and, ultimately, a more correct implementation.