Backfill architecture decision records (#494)
* chore(docs): backfill architecture decision records

I wanted to capture a new architectural/process decision we discussed
today, and found myself wanting to capture some past pivotal decisions
as well. I've been using architectural decision records (ADRs) in other
contexts for a bit now and have been seeing returns on that, so thought
we might give that a go here.

The important new decision revolves around managing token changes in a
distributed team setting and how to avoid issues such as broken styling
due to unexpectedly missing tokens. This proposal is currently in
"Proposed" status, so will need to update it to "Accepted" before
merging.

* fix(docs): move table of contents to index.md and create README

* fix(docs): remove duplicate record links

* fix: trigger new build