fix(storybook): update composed storybook config
In #543 I broke the composed Storybook, as I forgot to update its config
alongside the React and WebComponent Storybooks. I think this improves
things to the point that I'd like to merge it and try the build, but I
wasn't able to get the Storybooks to cooperate locally so it may still
be broken, ultimately.