feat(icon): fix aria-hidden logic
Because aria-hidden is a boolean attribute, it should be set to true
when the icon is hidden, and not exist at all when it is visible. Lit's
nothing attribute enables this behavior:
https://lit.dev/docs/templates/expressions/#removing-attribute This
also updates the logic to add aria-hidden if there is not description
or a11y-title to preserve the current behavior until description is
removed.