Add a11y properties to icon component (#646)
* feat(icon): add a11y properties to icon component

Added both a11y-label and a11y-hidden properties to the icon component.
Also marks the previous icon property description as deprecated. This is
in preparation of the v14 release which will remove the description
propererty and add an error if either a11y-label or a11y-hidden
are not provided.

* feat(icon): use a11y-title instead of a11y-label

The most accesssible way to label an svg icon is to use a title
element in the svg itself. This is better than using aria-label
which should be reserved for cases where an accessible name can
not be derived. See this page for more details:
https://developer.mozilla.org/en-US/docs/Web/SVG/Element/title

* feat(icon): fix aria-hidden logic

Because aria-hidden is a boolean attribute, it should be set to true
when the icon is hidden, and not exist at all when it is visible. Lit's
nothing attribute enables this behavior:
https://lit.dev/docs/templates/expressions/#removing-attribute This
also updates the logic to add aria-hidden if there is not description
or a11y-title to preserve the current behavior until description is
removed.

* fix(icon): fall back to description for title

If the title is not provided, fall back to the description. Previously,
it only fell back if the a11y-title was not provided, not an empty
string. This made the storybook stories fail.

* docs(icon): use new a11y attributes in storybook

* fix(icon): correct attribute name in changeset

* fix(icon): Fix typo in deprecation warning

Co-authored-by: Dane Hillard <github@danehillard.com>

* fix(icon): Update description deprecation text

Co-authored-by: Mat Harris <hurris88@gmail.com>

* docs(icon): Update new a11yTitle property definition

Co-authored-by: Mat Harris <hurris88@gmail.com>

* docs(icon): update title example in storybook

* feat(a11y): add global types for aria values

As we add a11y-* attributes as part of our accessibility work, we need
a global place to store shared types that multiple components can use.

* feat(icon): update a11yHidden to be a string

Because the end result of adding this attribute is setting a string
value, it makes sense to have the prop be a string. The end result is
that aria-hidden is either set to the string true, or not set at all.

* fix(icon): update a11yHidden false logic

---------

Co-authored-by: Dane Hillard <github@danehillard.com>
Co-authored-by: Mat Harris <hurris88@gmail.com>