fix(icon): update icon usage to new a11y syntax
Recent updates to the icon component were made which emit a
console log when deprecated attributes are used. This updates
all the internal usages of the icon component in other Pharos
components, storybook stories, and the pharos site to use the new
syntax so that the console log is no longer emitted in places that
would confuse our end users.