Update deprecated a11y attributes in button component (#666)
* fix(button): update test to use a11y-pressed

* fix(toggle-button): update to use a11y-pressed

The pressed attribute is deprecated in favor of a11y-pressed. This
updates the toggle button group to use the new attribute and stop
console logging deprecation warnings when it is used.

* fix(button): update label attribute to a11y-label

Recently the label attribute was deprecated in favor of a11y-label. This
updtes all the button uses internal to the library to use the new
attribute. This prevents confusing warnings in the console for the end
users.

* chore: add changeset