fix(storybook): update Babel configuration
The React Storybook is failing to build because when Storybook attempts
to build, it looks for a Babel configuration, and for story files it's
ascending and finding the root `babel.config.js` file and using only its
configuration. Namely, this doesn't include the React preset, and the
build fails to understand JSX syntax as a result.

Storybook offers a `babel` function for manipulating the Babel
configuration as a preferred approach, so this change moves the desired
configuration to the Storybook config and deletes the root Babel config
file, which was added mainly to appease the build earlier on.

- Also bumped Storybook + addons to the latest v7 release.
- Also added Netlify status badges to the README so these things go
  unnoticed less often.