chore(deps): update @custom-elements-manifest/analyzer (#723)
When I did the updates in #716, I had to pin this package to `0.9.0` to
keep things working without looking too deeply into the reason.

After that got merged and with some fresh cognitive capacity, I looked
into this and ultimately filed open-wc/custom-elements-manifest#242,
where the team was kind enough to guide me through what changed and what
the expectations are going forward for this format.

This change stops relying on the presence of `resolveInitializer`, which
is a leaky abstraction, and instead checks if there are any matching
variable values directly. I verified the output is identical, save for
the ordering of the exports in the built
`packages/pharos/src/react-components/index.ts` file, which is
definitely acceptable; I think they're even in alphabetical order, now.