fix: add individual component exports
We recently added an exports key to the package.json, which made
it so that components could only be imported from the root of the
package. In some situations this is not desired as it makes it more
difficult to share components between projects using module sharing
and causes bloat in the bundle. This commit adds back the individual
component exports.