docs: fix storybook link from pharos site
We host our storybook instance at pharos.jstor.org/storybook/, and the
trailing slash is required. We are linking to it using the Gatsby navigate
command, which was stripping the trailing slash. Stripping the trailing slash
is a global setting in Gatsby, this adds a redirect in gatsby-node.js to
handle this specific case.