Add Pill Component (#1066)
* feat(pill): add basic component structure

* feat(pill): add size options

* feat(pill): add dismissible pill

* feat(pill): add color attribute

* feat(pill): add primary and secondary variant

* feat(pill): add ability to pass in an icon to be displayed

* fix(pill): style cleanup to make dismissible and normal pills look the same

* test(pill): adjust timeout to avoid flakiness

* feat(pill): truncate text using ellipsis

* feat(color): add glacier-blue-30 tint

* feat(pill): add preset styling

* fix(pill): format new tokens

* fix(color): glacier-blue-30 hsl values had been truncated when I converted them, these are correct

* fix(pill): pass disabled attribute through to button when dismissible

* fix(pill): add separate text and border color tokens, make preset-1 text black to pass a11y

* fix(pill): adjust colors to pass WCAG contrast requirements

* fix(pill): adjust box-sizing of pill

* fix(pill): add icon container with flex

This prevents the icon from shrinking when
the pill is dismissible and has a set width.

* docs(pill): clean up storybook stories

* test(pill): add tests for presets

* fix(pill): set max-width on host

This will cause the pill to truncate when it exceeds
 the width of its container.

* docs(pill): update truncation and preset stories

* fix(pill): clean up code and comments

* fix(pill): fix icon attribute annotation

* docs(pill): setup controls and rearrange stories

* docs(pill): add react stories

* docs(pill): make truncation container resizable

* refactor(pill): remove unused PharosIcon element definition

* refactor(pill): tighten up types

* docs(pill): fix story types

* test(pill): use waitUntil for icon loading

This should make the tests less flaky, as they won't
rely on arbitrary timeouts.

* chore(pill): add changeset

* Update changelog to be more specific

Co-authored-by: Dane Hillard <github@danehillard.com>

* fix(pill): vertically align icon and text pills

Without this, when a pill with an icon is used next
to a pill that is only text, the icon pill will
render slightly higher than the text-only pill.

* fix(pill): update how button aria-label is set

Before it was querying the slotted text which
was causing it to be undefined because it hadn't
been rendered yet. Now it uses the component's
textContent which is available immediately.

* fix(pill): adjust padding of pill with an icon

With equal padding, the visual weight of the pill component is off
when there is an icon present. This commit adjusts the padding on the
right side of the pill when there is an icon to better balance the
visual weight.

* fix(pill): adjust dismissible padding

Similar to the last commit, this adds some extra padding to the left of
dismissible pills to visually balance the close icon. It also doesn't add
extra padding when both an icon and the close button are present, since
the visual weights are already balanced.

* docs(pill): add dismissible icon example to sizes story

* Copy edits for stories from code review

Co-authored-by: Dane Hillard <github@danehillard.com>

* Add pharos.jstor.org documentation for the new Pill component (#1084)

* docs(pill): add pill page to pharos site

* Copy edits from code review

Co-authored-by: Dane Hillard <github@danehillard.com>

* Reword example for using icons with pills

Clarify icon usage consistency in pill documentation.

---------

Co-authored-by: Dane Hillard <github@danehillard.com>

---------

Co-authored-by: Dane Hillard <github@danehillard.com>