fix(image-card): update collection error state
In #1145 we removed the display:grid from the image card component to
fix issues with images displaying in Firefox. However, this caused the
collection image card error to overflow its container. This update
restores the display:grid to the collection image card variant, and
fixes the Firefox issue by setting grid-area on the svg element.